Move SystemExecutionState to test_utils.
